refactor: move execution context into care.
