+++ rewrite-gradle/src/main/java/org/openrewrite/gradle/UpdateGradleWrapper.java
@@ -71,7 +71,7 @@ public class UpdateGradleWrapper extends Recipe {
     @Override
     public Validated validate(ExecutionContext ctx) {
         if (gradleWrapper == null) {
-            org.openrewrite.ipc.http.HttpSender httpSender = GradleExecutionContextView.view(ctx).getHttpSender();
+            org.openrewrite.ipc.http.HttpSender httpSender = HttpSenderExecutionContextView.view(ctx).getHttpSender();
             gradleWrapper = super.validate().and(GradleWrapper.validate(version, distribution, httpSender));
         }
         return gradleWrapper;
